requirements.txt
Django==1.8.1
PyYAML==3.11
django-rest-swagger==0.2.9
djangorestframework==3.1.2
psycopg2==2.6
python-gcm==0.1.5
wsgiref==0.1.2
